Dynamic threads_array in the heap, i.e. massive parallel multi-threading
For what it's worth, this branch implements up to 16.000 threads (just an
arbitrary limit) by allocating threads_array on the heap.
I do not want to realloc this array on demand when a new task is created,
I rather want it to be at the same location.

What I'm not sure is if the tasks and threads need to be killed at interp
teardown at all, or just be freed. Cancelling or terminating a thread is
not really friendly. The current code passes all tests.