Seperate thread_kill from thread_destroy
Rename Parrot_thread_destroy to Parrot_thread_destroy_all
Add an earlier Parrot_thread_kill_all step, which kills the
foreign tasks, and use the later Parrot_thread_destroy_all to
cancel the threads, kill its interpreters and free the threads_array.

I'm not sure if the scenario is sound and right but the tests pass.